+++ src/components/date-range/date-range.component.tsx
@@ -28,7 +28,6 @@ import DateRangeContext, {
   SetInputRefMapValue,
 } from "./__internal__/date-range.context";
 import NewValidationContext from "../carbon-provider/__internal__/new-validation.context";
-import Logger from "../../__internal__/utils/logger";
 
 interface DateInputValue {
   formattedValue: string;
@@ -115,11 +114,6 @@ export interface DateRangeProps
   tooltipPosition?: "top" | "bottom" | "left" | "right";
   /** Flag to configure component as mandatory. */
   required?: boolean;
-  /**
-   * [Legacy] Flag to configure component as optional.
-   * @deprecated If the value of this component is not required, use the `required` prop and set it to false instead.
-   */
-  isOptional?: boolean;
   /** Date format string to be applied to the date inputs */
   dateFormatOverride?: string;
   /** Prop to specify the aria-label attribute of the start date picker */
@@ -133,8 +127,6 @@ export interface DateRangeProps
   /** Render the ValidationMessage above the Date inputs when validationRedesignOptIn flag is set */
   validationMessagePositionTop?: boolean;
 }
-
-let deprecateOptionalWarnTriggered = false;
 
 export const DateRange = ({
   endDateProps = {},
@@ -150,7 +142,6 @@ export const DateRange = ({
   startRef,
   endRef,
   required,
-  isOptional,
   dateFormatOverride: dateFormatOverrideProp,
   datePickerStartAriaLabel,
   datePickerStartAriaLabelledBy,
@@ -159,12 +150,6 @@ export const DateRange = ({
   validationMessagePositionTop = true,
   ...rest
 }: DateRangeProps) => {
-  if (!deprecateOptionalWarnTriggered && isOptional) {
-    deprecateOptionalWarnTriggered = true;
-    Logger.deprecate(
-      "`isOptional` is deprecated in DateRange and support will soon be removed. If the value of this component is not required, use the `required` prop and set it to false instead.",
-    );
-  }
   const { validationRedesignOptIn } = useContext(NewValidationContext);
   const labelsInlineWithNewValidation = validationRedesignOptIn
     ? false
@@ -383,7 +368,6 @@ export const DateRange = ({
         handleOnKeyDown(ev, propsKey),
       ...props,
       required,
-      isOptional,
     };
   };