+++ src/components/date-range/date-range.stories.tsx
@@ -194,27 +194,6 @@ export const Required: Story = () => {
 };
 Required.storyName = "Required";
 
-export const IsOptional: Story = () => {
-  const [state, setState] = useState(["01/10/2016", "30/10/2016"]);
-  const handleChange = (ev: DateRangeChangeEvent) => {
-    const newValue = [
-      ev.target.value[0].formattedValue,
-      ev.target.value[1].formattedValue,
-    ];
-    setState(newValue);
-  };
-  return (
-    <DateRange
-      startLabel="Start"
-      endLabel="End"
-      onChange={handleChange}
-      value={state}
-      isOptional
-    />
-  );
-};
-IsOptional.storyName = "IsOptional";
-
 export const LocaleFormatOverrideExampleImplementation: Story = ({
   ...args
 }) => {