+++ src/components/grouped-character/grouped-character.stories.tsx
@@ -207,26 +207,6 @@ export const Required: Story = () => {
 };
 Required.storyName = "Required";
 
-export const IsOptional: Story = () => {
-  const [state, setState] = useState("1231231");
-
-  const setValue = ({ target }: CustomEvent) => {
-    setState(target.value.rawValue);
-  };
-
-  return (
-    <GroupedCharacter
-      label="GroupedCharacter"
-      value={state}
-      onChange={setValue}
-      groups={[2, 2, 3]}
-      separator="-"
-      isOptional
-    />
-  );
-};
-IsOptional.storyName = "IsOptional";
-
 export const LabelAlign: Story = () => {
   const [state, setState] = useState({
     right: "1231231",