+++ src/components/textarea/textarea.test.tsx
@@ -17,27 +17,6 @@ jest.mock("../../__internal__/utils/helpers/guid");
 const mockedGuid = "guid-12345";
 (guid as jest.MockedFunction<typeof guid>).mockImplementation(() => mockedGuid);
 
-test("should display deprecation warning once when rendered as optional", () => {
-  const loggerSpy = jest.spyOn(Logger, "deprecate");
-
-  render(
-    <>
-      <Textarea name="my-textarea" onChange={() => {}} isOptional />
-      <Textarea name="my-textarea" onChange={() => {}} isOptional />
-    </>,
-  );
-
-  // Ensure the deprecation warning is logged only once
-  expect(loggerSpy).toHaveBeenCalledTimes(1);
-
-  expect(loggerSpy).toHaveBeenNthCalledWith(
-    1,
-    "`isOptional` is deprecated in TextArea and support will soon be removed. If the value of this component is not required, use the `required` prop and set it to false instead.",
-  );
-
-  loggerSpy.mockRestore();
-});
-
 test("should display deprecation warning once when rendered as uncontrolled", () => {
   const loggerSpy = jest.spyOn(Logger, "deprecate");
 
@@ -45,20 +24,6 @@ test("should display deprecation warning once when rendered as uncontrolled", ()
 
   expect(loggerSpy).toHaveBeenCalledWith(
     "Uncontrolled behaviour in `Textarea` is deprecated and support will soon be removed. Please make sure all your inputs are controlled.",
-  );
-
-  expect(loggerSpy).toHaveBeenCalledTimes(1);
-
-  loggerSpy.mockRestore();
-});
-
-test("should display deprecation warning once for `ariaDescribedBy`", () => {
-  const loggerSpy = jest.spyOn(Logger, "deprecate");
-
-  render(<Textarea onChange={() => {}} ariaDescribedBy="test" />);
-
-  expect(loggerSpy).toHaveBeenCalledWith(
-    "The `ariaDescribedBy` prop in `Textarea` is deprecated and will soon be removed, please use `aria-describedby` instead.",
   );
 
   expect(loggerSpy).toHaveBeenCalledTimes(1);
@@ -375,20 +340,6 @@ test("appends the provided `aria-describedby` to the accessible description", ()
     <>
       <p id="test">description</p>
       <Textarea inputHint="hint text" aria-describedby="test" />
-    </>
-  );
-  render(<Component />);
-
-  expect(screen.getByRole("textbox")).toHaveAccessibleDescription(
-    "hint text description",
-  );
-});
-
-test("appends the provided `ariaDescribedBy` to the accessible description", () => {
-  const Component = () => (
-    <>
-      <p id="test">description</p>
-      <Textarea inputHint="hint text" ariaDescribedBy="test" />
     </>
   );
   render(<Component />);