+++ src/components/vertical-menu/responsive-vertical-menu/responsive-vertical-menu.context.tsx
@@ -10,6 +10,7 @@ import React, {
 } from "react";
 
 import { IconType } from "../../icon/icon-type";
+import Logger from "../../../__internal__/utils/logger";
 
 export interface ResponsiveVerticalMenuButtonItem {
   children?: ReactNode;
@@ -29,10 +30,16 @@ export interface MenuContextType {
   menuRef: RefObject<HTMLUListElement>;
   reducedMotion?: boolean;
   responsiveMode?: boolean;
+  left: string;
+  top: string;
+  width?: string;
+  height?: string;
   setActive: Dispatch<SetStateAction<boolean>>; // This allows both value and function
   setActiveMenuItem: (item: ResponsiveVerticalMenuButtonItem | null) => void;
   setReducedMotion?: (reducedMotion: boolean) => void;
   setResponsiveMode?: (responsiveMode: boolean) => void;
+  setLeft: (left: string) => void;
+  setTop: (top: string) => void;
 }
 
 export const ResponsiveVerticalMenuContext =
@@ -50,11 +57,29 @@ export const useResponsiveVerticalMenu = () => {
 
 export interface ResponsiveVerticalMenuProviderProps {
   children: ReactNode;
+  /** @private @internal @ignore */
+  width?: string;
+  /** @private @internal @ignore */
+  height?: string;
 }
+
+let deprecatedWarning = false;
 
 export const ResponsiveVerticalMenuProvider = ({
   children,
+  width,
+  height,
 }: ResponsiveVerticalMenuProviderProps) => {
+  const externalContext = useContext(ResponsiveVerticalMenuContext);
+  const hasExternalContext = externalContext !== null;
+
+  if (!deprecatedWarning && hasExternalContext) {
+    deprecatedWarning = true;
+    Logger.deprecate(
+      "`ResponsiveVerticalMenuProvider` is deprecated and no longer needed for `ResponsiveVerticalMenu`. You can use `ResponsiveVerticalMenu` directly without wrapping it in a provider.",
+    );
+  }
+
   const [active, setActive] = useState<boolean>(false);
   const [activeMenuItem, setActiveMenuItem] =
     useState<ResponsiveVerticalMenuButtonItem | null>(null);
@@ -64,6 +89,9 @@ export const ResponsiveVerticalMenuProvider = ({
   const menuRef = useRef<HTMLUListElement>(null);
   const [responsiveMode, setResponsiveMode] = useState<boolean>(false);
   const [reducedMotion, setReducedMotion] = useState<boolean>(false);
+
+  const [left, setLeft] = useState("auto");
+  const [top, setTop] = useState("auto");
 
   return (
     <ResponsiveVerticalMenuContext.Provider
@@ -75,10 +103,16 @@ export const ResponsiveVerticalMenuProvider = ({
         menuRef,
         reducedMotion,
         responsiveMode,
+        left,
+        top,
+        width,
+        height,
         setActive,
         setActiveMenuItem,
         setReducedMotion,
         setResponsiveMode,
+        setLeft,
+        setTop,
       }}
     >
       {children}