+++ src/locales/locale.ts
@@ -74,9 +74,6 @@ interface Locale {
   heading: {
     backLinkAriaLabel: () => string;
   };
-  label: {
-    optional: () => string;
-  };
   link: {
     skipLinkLabel: () => string;
   };